[SERCOM] Regenerated demo for sercom_spi_spi_eeprom_write_read_sam_d20_xpro MH3-27835

+++ apps/sercom/spi/spi_eeprom_write_read/firmware/src/config/sam_d20_xpro/peripheral/sercom/spim/plib_sercom1_spi.c
@@ -164,7 +164,7 @@ bool SERCOM1_SPI_TransferSetup(SPI_TRANSFER_SETUP *setup, uint32_t spiSourceCloc
         if((baudValue > 0) & (baudValue <= 255))
         {
             /* Selection of the Clock Polarity and Clock Phase */
-            SERCOM1_REGS->SPIM.SERCOM_CTRLA |= setup->clockPolarity | setup->clockPhase;
+            SERCOM1_REGS->SPIM.SERCOM_CTRLA |= (uint32_t)setup->clockPolarity | (uint32_t)setup->clockPhase;
 
             /* Selection of the Baud Value */
             SERCOM1_REGS->SPIM.SERCOM_BAUD = baudValue;
